kinesis-forwarder-py/requirements.txt
aws_kinesis_agg
avro-python3